% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prob.R
\name{calc_post}
\alias{calc_post}
\title{Posterior probability of success}
\usage{
calc_post(num, rate, prob)
}
\arguments{
\item{num}{Number of subjects.}

\item{rate}{underlying probability of success.}

\item{prob}{Threshold for posterior probability.}
}
\value{
A data frame (tibble) with the following columns:
\itemize{\item \strong{n}: Number of subjects
\item \strong{res}: Number of responders
\item \strong{prob}: Achieved posterior probability of success
}
}
\description{
The main function to calculate the number of success required to ensure P(post>p1)>=p2
}
\details{
\code{\link[=calc_post]{calc_post()}} returns number successes required such that the posterior probability of achieving a success rate p1
exceeds p2
}
\examples{
# Generate number of successes for 80 subjects.  Rate, p1, is 0.3.
# Posterior probability threshold ranges from
# 0.5 to 0.8.
p1 <- 0.3
p2 <- seq(0.5, 0.8, 0.1)
post_prob <- calc_post(80, p1, p2)
}
