% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_prob.R
\name{calc_post}
\alias{calc_post}
\title{Posterior probability of success}
\usage{
calc_post(num, rate, prob)
}
\arguments{
\item{num}{Number of subjects.}

\item{rate}{Threshold for probability of success, \eqn{p_1}.}

\item{prob}{Threshold for posterior probability, \eqn{p_2}.}
}
\value{
A data frame (tibble) with the following columns:
\itemize{\item \strong{n}: Number of subjects
\item \strong{rate}: \eqn{p_1}
\item \strong{prob}: \eqn{p_2}
\item \strong{res}: Number of responders
\item \strong{post_prob}: Achieved posterior probability of success, \eqn{p_2}.
}
}
\description{
The main function to calculate the number of success required, given number of subjects \eqn{n},
to satisfy conditions for the posterior probability of the underlying success rate, p_post.
}
\details{
\code{\link[=calc_post]{calc_post()}} returns number successes required such that \eqn{P({\sf p\_post}>p_1) \ge p_2}.
}
\examples{
# Generate number of successes for up to 80 subjects.  Rate, p1, is 0.3.
# Posterior probability threshold ranges from 0.5 to 0.8.
p1 <- 0.3
p2 <- seq(0.5, 0.8, 0.1)
post_prob <- calc_post(80, p1, p2)
post_prob
}
